refactor: centralize impersonation storage and validation logic in src/utils/impersonation.ts
